docs: add getting started documentation
I created the software requirements documentation to inform developers
on what software is required to be installed in order to develop for
Time. I updated README.md with a link to the software requirements
documentation and instructions on how to clone the Git repository
locally and set it up for local development.