style: add copyright and license text
I updated all of the source code files and added a standard file header
containing the copyright statement and the text of the license
agreement. The purpose of including the text of the license agreement is
to ensure that the license terms are communicated if individual source
files are redistributed separately from the rest of the product source
code to inform the recipients of the license terms covering the source
code.