ci: fix container build error
I fixed a problem with the build of Time in the container image. It was
using the same executables built for Linux, but I forgot that certain
library dependencies and capabilities are not supported in the
containers. I used Claude Sonnet 4.5 to fix the builds so that different
Linux executables are created for the container image.