ci: fix missing execute permission
The time programs built for Linux and Apple macOS were not carrying
their execute flag in the distribution archives. I updated release.yml
to explicitly set the execute flag before archiving, and then to ensure
that the archive carried the execute flag so that the time program is
executable after extracting it from the archive.