ci: update version tagging
I updated the `docker-container.yml` workflow and the `release.yml`
workflow to create additional tags. For the container image, the image
will be tagged with the full semantic version (major.minor.patch). For
the GitHub release, the release will also be tagged with the `v<major>`
and `v<major>.<minor>` tags.