ci: create GitHub release
I created a GitHub Actions workflow that will execute on release
branches. The workflow will generate a draft release for a release pull
request and will build the changelog for the release. The GitHub release
will be updated as new changes are pushed to the pull request. When the
pull request is closed, the release will be published and the version
will be tagged.

Closes: #1