ci: fix container image tagging
I fixed an issue with the container image not getting tagged with the
version number for the product. While the release is in the draft phase,
the tag for the container image is not being created and then the
container metadata step doesn't properly tag the container image. The
container image needs to be built again after the release is released so
that the tag exists and the right labels can be attached to the
container image.