ci: fix container image tagging
I fixed `release.yml` to fix the tagging of the container image. The
container image isn't getting tagged with the version number labels
because the metadata step needs to run for a tag event, not the event of
the PR being closed. I modified the workflow to build and publish the
release container image when the version tag is created.