ci: fix container image tagging
I worked with GitHub Copilot to try to fix the version tagging of the
container image when generating releases. The previous fix didn't fix
the problem because the workflow to build and publish the tagged
container image did not execute. We made some changes to the workflow
and split out the `publish-docker.yml` workflow to do the tagging when
the version tag is created in the GitHub repository.