ci: fix container image tagging
The `publish-docker.yml` workflow didn't run for the 0.0.5 release
again. This is probably because the change was made in the same commit
where the tag was created, so GitHub Actions didn't have time to process
the workflow and register the trigger. This fix adds the
`workflow_dispatch` event handler so that it can be run manually.