feat: create an activity
I implemented the `activity add` command to add an activity to the
user's Activity Inventory. I refactored the CLI to move the pomodoro
command under the `pomodoro start` subcommand.

I updated README.md with instructions on how to create an activity and
then how to use the new `pomodoro start` command to start a pomodoro and
link the pomodoro to an activity.