feat: associate activity with pomodoro
I updated the `pomodoro start` command to accept the identifier of an
activity that the pomodoro is working on. If the activity is not
provided, the command will attempt to read the identifier from stdin.
This will create an association in the database between the activity and
the pomodoro for tracking later.

Closes: #12