feat: select an activity
I implemented a new feature for selecting an activity. The `activity
select` command will display the activities in the Activity Inventory in
a list. When the user selects an activity, the program will output the
activity's identifier to stdout. This will allow the `activity select`
command to be connected to another command using a pipe operator to
send the activity identifier to the input of another command.

I refactored the existing command code to move the text UI logic and
command support code into the cli package. This made more sense to keep
the command and its immediate implementation closer together than to
put them in separate packages.